+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- numerai-quant API
+
+
 ## [2.2.4] - 2020-05-11
 - Remove required model_id annotation for submissions status lookups so that None can be passed
 - Use consistent modelId in query spec